move DiffusionModel from experimental to networks
Stabilizes the DiffusionModel class. A deprecation warning for the
DiffusionModel class in the experimental module was added.
