+++ bayesflow/networks/__init__.py
@@ -7,6 +7,7 @@ The module features inference networks (IN), summary networks (SN), as well as g
 from .consistency_models import ConsistencyModel
 from .coupling_flow import CouplingFlow
 from .deep_set import DeepSet
+from .diffusion_model import DiffusionModel
 from .flow_matching import FlowMatching
 from .inference_network import InferenceNetwork
 from .point_inference_network import PointInferenceNetwork
@@ -19,4 +20,4 @@ from .transformers import SetTransformer, TimeSeriesTransformer, FusionTransform
 
 from ..utils._docs import _add_imports_to_all
 
-_add_imports_to_all(include_modules=[])
+_add_imports_to_all(include_modules=["diffusion_model"])
