+++ bayesflow/networks/diffusion_model/schedules/cosine_noise_schedule.py
