+++ bayesflow/networks/diffusion_model/schedules/edm_noise_schedule.py
